Poprawienie calego kodu

+++ src/main/java/pl/javastart/todoapp/service/TaskService.java
@@ -1,6 +1,7 @@
 package pl.javastart.todoapp.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import pl.javastart.todoapp.model.Task;
 import pl.javastart.todoapp.repository.TaskRepository;
@@ -9,24 +10,29 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
+@Slf4j
+@RequiredArgsConstructor
 public class TaskService {
-    @Autowired
-    private TaskRepository taskRepository;
 
-    public List<Task> getAllTasks(){
+    private final TaskRepository taskRepository;
+
+    public List<Task> getAllTasks() {
         return taskRepository.findAll();
     }
 
-    public Optional<Task> getById(Long id){
+    public Optional<Task> getById(Long id) {
         return taskRepository.findById(id);
     }
 
-    public Task createTask(Task task){
-        return taskRepository.save(task);
+    public Task createTask(Task task) {
+        Task createdTask = taskRepository.save(task);
+        log.info("Created new task: {}", task.getTitle());
+        log.error("Something wrong!");
+        return createdTask;
     }
 
-    public Optional<Task> updateTask(Long id, Task updatedTask){
-        return taskRepository.findById(id).map(existingTask ->{
+    public Optional<Task> updateTask(Long id, Task updatedTask) {
+        return taskRepository.findById(id).map(existingTask -> {
             existingTask.setTitle(updatedTask.getTitle());
             existingTask.setDescription(updatedTask.getDescription());
             existingTask.setStatus(updatedTask.getStatus());
@@ -34,8 +40,9 @@ public class TaskService {
         });
     }
 
-    public boolean deleteTask(Long id){
-        if(taskRepository.existsById(id)){
+    public boolean deleteTask(Long id) {
+        if (id == null) return false;
+        if (taskRepository.existsById(id)) {
             taskRepository.deleteById(id);
             return true;
         }
